+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * feat: add `ProjectTransferInvitation` interface
 * feat: add project transfer invitations endpoints (index, show, create, accept, decline, cancel)
 * feat: add `in_free_trial` to `AppOwner` interface
+* feat: add events endpoint on `Project`
 
 ## 0.16.1